+++ plugins/modules/dhcp.py
@@ -1,14 +1,14 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 #
-# Copyright: (c) 2020-2022, Men&Mice
+# Copyright: (c) 2020-2023, Men&Mice
 # GNU General Public License v3.0
 # see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt
 """Ansible DHCP reservation module.
 
 Part of the Men&Mice Ansible integration
 
-Module to manage DHCP reservations in the Men&Mice Suite
+Module to manage DHCP reservations in the Micetro
   - Set or release a DHCP reservation
 """
 
@@ -27,12 +27,12 @@ from ansible_collections.menandmice.ansible_micetro.plugins.module_utils.micetro
 
 DOCUMENTATION = r"""
   module: dhcp
-  short_description: Manage DHCP reservations on the Men&Mice Suite
+  short_description: Manage DHCP reservations on the Micetro
   author:
     - Ton Kersten <t.kersten@atcomputing.nl> for Men&Mice
   version_added: "2.7"
   description:
-    - Manage DHCP reservations on the Men&Mice Suite
+    - Manage DHCP reservations on the Micetro
   notes:
     - When in check mode, this module pretends to have done things
       and returns C(changed = True).
@@ -81,7 +81,7 @@ DOCUMENTATION = r"""
       type: bool
       required: False
     mm_provider:
-      description: Definition of the Men&Mice suite API mm_provider.
+      description: Definition of the Micetro API mm_provider.
       type: dict
       required: True
       suboptions:
@@ -108,7 +108,7 @@ EXAMPLES = r"""
     ipaddress: 172.16.17.8
     macaddress: 44:55:66:77:88:99
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost