+++ plugins/modules/dnsrecord.py
@@ -1,14 +1,14 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 #
-# Copyright: (c) 2020-2022, Men&Mice
+# Copyright: (c) 2020-2023, Men&Mice
 # GNU General Public License v3.0
 # see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt
 """Ansible DNS Record Management module.
 
 Part of the Men&Mice Ansible integration
 
-Module to manage DNS entries for IP addresses in the Men&Mice Suite
+Module to manage DNS entries for IP addresses in the Micetro
 """
 
 from __future__ import absolute_import, division, print_function
@@ -24,12 +24,12 @@ from ansible_collections.menandmice.ansible_micetro.plugins.module_utils.micetro
 
 DOCUMENTATION = r"""
   module: dnsrecord
-  short_description: Manage DNS records in the Men&Mice Suite
+  short_description: Manage DNS records in the Micetro
   author:
     - Ton Kersten <t.kersten@atcomputing.nl> for Men&Mice
   version_added: "2.7"
   description:
-    - Manage DNS records in the Men&Mice Suite.
+    - Manage DNS records in the Micetro.
   notes:
     - When in check mode, this module pretends to have done things
       and returns C(changed = True).
@@ -102,7 +102,7 @@ DOCUMENTATION = r"""
       required: False
       default: 0
     mm_provider:
-      description: Definition of the Men&Mice suite API mm_provider.
+      description: Definition of the Micetro API mm_provider.
       type: dict
       required: True
       suboptions:
@@ -130,7 +130,7 @@ EXAMPLES = r"""
     rrtype: A
     dnszone: example.net.
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost
@@ -143,7 +143,7 @@ EXAMPLES = r"""
     rrtype: PTR
     dnszone: "17.16.172.in-addr.arpa."
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost
@@ -157,7 +157,7 @@ EXAMPLES = r"""
     data: "10 ringo"
     ttl: 86400
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost