+++ plugins/modules/ipprops.py
@@ -1,14 +1,14 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 #
-# Copyright: (c) 2020-2022, Men&Mice
+# Copyright: (c) 2020-2023, Men&Mice
 # GNU General Public License v3.0
 # see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt
 """Ansible IP properties address module.
 
 Part of the Men&Mice Ansible integration
 
-Module to set properties on an IP addresses in the Men&Mice Suite
+Module to set properties on an IP addresses in the Micetro
 """
 
 from __future__ import absolute_import, division, print_function
@@ -24,12 +24,12 @@ from ansible_collections.menandmice.ansible_micetro.plugins.module_utils.micetro
 
 DOCUMENTATION = r"""
   module: ipprops
-  short_description: Set properties on an IP address in the Men&Mice Suite
+  short_description: Set properties on an IP address in the Micetro
   author:
     - Ton Kersten <t.kersten@atcomputing.nl> for Men&Mice
   version_added: "2.7"
   description:
-    - Set properties on an IP address in the Men&Mice Suite
+    - Set properties on an IP address in the Micetro
     - This can be properties as custom properties, claim and so on
   notes:
     - When in check mode, this module pretends to have done things
@@ -59,7 +59,7 @@ DOCUMENTATION = r"""
       type: dict
       required: True
     mm_provider:
-      description: Definition of the Men&Mice suite API mm_provider.
+      description: Definition of the Micetro API mm_provider.
       type: dict
       required: True
       suboptions:
@@ -87,7 +87,7 @@ EXAMPLES = r"""
       claimed: false
       location: London
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost