+++ plugins/modules/props.py
@@ -1,14 +1,14 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 #
-# Copyright: (c) 2020-2022, Men&Mice
+# Copyright: (c) 2020-2023, Men&Mice
 # GNU General Public License v3.0
 # see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt
 """Ansible Custom Properties Management module.
 
 Part of the Men&Mice Ansible integration
 
-Module to manage custom properties in the Men&Mice Suite
+Module to manage custom properties in the Micetro
 """
 
 from __future__ import absolute_import, division, print_function
@@ -23,12 +23,12 @@ from ansible_collections.menandmice.ansible_micetro.plugins.module_utils.micetro
 
 DOCUMENTATION = r"""
   module: props
-  short_description: Manage custom properties in the Men&Mice Suite
+  short_description: Manage custom properties in the Micetro
   author:
     - Ton Kersten <t.kersten@atcomputing.nl> for Men&Mice
   version_added: "2.7"
   description:
-    - Manage custom properties in the Men&Mice Suite
+    - Manage custom properties in the Micetro
   notes:
     - When in check mode, this module pretends to have done things
       and returns C(changed = True).
@@ -103,7 +103,7 @@ DOCUMENTATION = r"""
       type: bool
       default: False
     mm_provider:
-      description: Definition of the Men&Mice suite API mm_provider.
+      description: Definition of the Micetro API mm_provider.
       type: dict
       required: True
       suboptions:
@@ -130,7 +130,7 @@ EXAMPLES = r"""
     proptype: text
     dest: zone
     mm_provider:
-      mm_url: http://mmsuite.example.net
+      mm_url: http://micetro.example.net
       mm_user: apiuser
       mm_password: apipasswd
   delegate_to: localhost